requirements.txt
python-binance>=1.0.19
websockets>=12.0
requests>=2.31.0
python-dotenv>=1.0.0
